fix: (fix) use specific rest endpoints for MCP
